// thy-translate/models/thy-translate.enum.ts
export enum ThyLangIdentifiers {
    English = 'en',
    French = 'fr',
    German = 'de',
    Swedish = 'sw',
    Norwegian = 'no',
    Dutch = 'nl'
}


export enum ThyLangFlags {
    English = 'flag-usa',
    French = 'flag-france',
    German = 'flag-germany',
    Swedish = 'flag-sweden',
    Norwegian = 'flag-norway',
    Dutch = 'flag-netherlands'
}