// thy-period/models/thy-period-unit.enum.ts
export enum ThyPeriodUnit {
    None,
    Minute = 2,
    Hour = 3,
    Day = 4,
    Week = 5,
    Month = 6,
    Year = 7,
    Decennary = 8
}

export enum ThyPeriodUnitStr {
    None = 'None',
    Hour = 'Hour',
    Day = 'Day',
    Week = 'Week',
    Month = 'Month',
    Year = 'Year',
    Decennary = 'Decennary'
}

export enum ThyPeriodUnitChar {
    None = 'N',
    Hour = 'H',
    Day = 'D',
    Week = 'W',
    Month = 'M',
    Year = 'Y',
    Decennary = 'D'
}

export enum ThyPeriodCurrentStr {
    None = 'None',
    Hour = 'CurrentHour',
    Day = 'CurrentDay',
    Week = 'CurrentWeek',
    Month = 'CurrentMonth',
    Year = 'CurrentYear',
    Decennary = 'CurrentDecennary'
}